net: phy: marvell: Implement downshift tunable
If the Ethernet cable only has two pairs, it cannot be used for 1Gbps.
If downshift is enabled, the PHY will retry a number of times to
establish a link, and if it fails it will downshift to 100Mbps.

This tunable can be used to enable/disable the feature, and to set the
number of retires before shifting down to a lower speed.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>