net: phy: marvell: Centralize setting polarity during config_init
Use a hw_feature flag to indicate if the PHY is limited to Fast
Ethernet. Use this to determine which function to use to configure
polarity during config_init. We now consistently set it for all PHYs.
At the same time, update the phydev idea of the polarity from unknown
to auto crossover.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>