net: dsa: mv88e6xxx: scratch registers and external MDIO pins
MV88E6352 and later switches support GPIO control through the "Scratch
& Misc" global2 register. Add accessors for these scratch
registers.

Two of the pins controlled this way on the mv88e6390 family are the
external MDIO pins. They can either by used as part of the MII
interface for port 0, GPIOs, or MDIO. Add a function to configure them
for MDIO, if possible.

Suggested-by: Russell King <rmk@armlinux.org.uk>
Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
Signed-off-by: Andrew Lunn <andrew@lunn.ch>