net: br: Fix igmp snooping offload with CONFIG_BRIDGE_VLAN_FILTERING
With CONFIG_BRIDGE_VLAN_FILTERING enabled, but the feature not enabled
via /sys/class/net/brX/bridge/vlan_filtering, mdb offloaded to the
kernel have the wrong VID.

When an interface is added to the bridge, switchdev is first used to
notify the hardware that a port has joined a bridge. This is
immediately followed by the default_pvid, 1, being added to the
interface via another switchdev call.

The bridge will then perform IGMP snooping, and offload an mdb entries
to the switch as needed. With vlan filtering disabled, the vid is left
as 0. This causes the switch to put the static mdb into the wrong
vlan, and it is not applied to frames.

If vlan filtering is disable, use the default_pvid, not 0.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>