net: dsa: mv88e6xxx: Enable external MDIO bus
The mv88e6390 family has an external MDIO bus. It can only be used if
the pins are not being used for MII on port 0, and the pins are not
configured for GPIO.

If the external MDIO bus is listed in the device tree, try to configure
the pins for MDIO usage.

Suggested-by: Russell King <rmk@armlinux.org.uk>
Signed-off-by: Andrew Lunn <andrew@lunn.ch>