ARM: dts: vf610-zii-dev-c: Add 10Gbps Ethernet port
Now that the external MDIO bus works, we can add the 10Gbps external
PHY to the device tree, which is connected to port 9 of the first
switch. Set the phy-mode to rxaui, although what mode it actually uses
depends on the link it negotiates with the peer. This is then
indicated via the adjust_link callback.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>