net: dsa: mv88e6xxx: Add SERDES stats for the 6390 family
The SERDES interfaces of the 6390 family have a number of statistic
counters. Add them to the statistics returned by ethtool -S. SERDES
statistics are not listed all the time. Some ports don't have a SERDES
interface. Depending on the mode of port 9 and 10, some of the lower
ports don't have a SERDES interface available. The SERDES may also not
be active unless the port is using SGMII, or for ports 9 and 10,
1000Base-X or higher modes.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>