ethtool: provide message level in GET_SETTINGS request
Add information about supported and enabled message levels to the
GET_SETTINGS reply when ETHTOOL_IM_SETTINGS_DEBUG flag is set in the
request.

Unlike in the ioctl interface, "message level" is called "message mask" as
it is in fact interpreted as a bit mask. It is put into a nested attribute
ETHTOOL_A_SETTINGS_DEBUG to allow future extensions (e.g. an actual
verbosity level).

Send notification in the same format as reply when message level is
modified using the ioctl interface (ETHTOOL_SMSGLVL command).

Signed-off-by: Michal Kubecek <mkubecek@suse.cz>