media: coda: pad bitstream to release coda from ring buffer empty condition
If the h.264 or MPEG-2 decoder hangs waiting for more input because we
estimated the prefetcher target incorrectly, write some dummy data into
the bitstream and kick the decoder to continue.

This works on h.264 by writing a filler NAL and on MPEG-2 by writing a
user data sequence of sufficient size.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>