media: coda: MPEG-2 end of stream handling
I assume flag 0x40000 is the end of stream signal of the MPEG-2 decoder
firmware. Once it is set, the firmware does not decode any further
frames, but it still continues to point out the remaining already
decoded frames for presentation. Once the last frame is reached, signal
it via the V4L2_BUF_FLAG_LAST mechanism.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>