net: phy: Add cable test support to state machine
Running a cable test is desruptive to normal operation of the PHY and
can take a 5 to 10 seconds to complete. The RTNL lock cannot be held
for this amount of time, and add a new state to the state machine for
running a cable test.

The driver is expected to implement three functions. The first is used
to start a cable test. Once the test has started, it should return.

The second function is called once per second, or on interrupt to
check if the cable test is complete, and to allow the PHY to report
the status.

The third function is not normally used. However, if the MAC driver
disconnects the PHY, or the systems suspends, it is necassary to abort
the cable test and place it into a state where normal operations can
be resumed, or the device suspend.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>