watchdog: rave-sp-wdt: let other restart handlers run
The RAVE SP watchdog restart handler does nothing but wait, as the
actual restart request is issued earlier in the reboot notifier.
Returning an error causes the watchdog core to return NOTIFY_BAD, and to
skip the other restart handlers. This causes restarts that skip the
reboot notifiers altogether (such as sysrq 'b' or emergency restarts
after an oops panic) to fail. Returning 0 allows the internal i.MX
watchdog restart handler to be executed as a fallback.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>