net: ethtool: Add Properties for cable test reports.
Add the attributes needed to report cable test results to userspace.
The reports are expected to be per twisted pair. A nested property per
pair can report the result of the cable test, and length of the cable
to any fault. More attributes can be added later for other types of
fault or measurement.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>