net: phy: Add helpers and attributes for amplitude graph
The amplitude graph needs to return the measured amplitude of the
reflected signal for each pair. Add a helper to store such a
measurement into the results skbuf. The size of the transmitted pulse
affects the size of the measured pulse. So add a helper to report the
pulse size.

For a 100m cable, and measurements at one meter interval, the default
size netlink skbuf is too small. Change to 8K.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>