ARM: dts: imx51-zii-rdu1: set parallel display encoder interface format
Currently the parallel to LVDS display bridge is not present in the DT and
won't be for a while until upstream has worked out how to support bridges
that change the bus format. Set the interface-pix-fmt which overrides the
bus format provided by the panel, so we can link the real LVDS panel
description to the parallel encoder. This is a bit of a hack, as the
property is deprecated, but it allows us to work on a proper solution,
which won't require backward incompatible DT changes.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>