media: coda: parse and replay H.264 headers
Store a copy of the H.264 SPS and PPS headers and reuse them if the
first buffer after a stream restart does not contain new headers.

This allows to support keyframe-only reverse playback as currently
implemented by GStreamer without having to enable SPS and PPS insertion
with every IDR frame.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>