power: supply: ucs1002: disable power if max current is 0
For some devices userspace needs the ability to completely cut the power
to the USB devices connected to the charge controller. An easy way to
achieve this is by allowing 0 as a valid max current and disable the
regulator in that case, as well as enable the regulator again if it's
in use and a non-0 max current is set.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>