Input: exc3000: add firmware update support
This change allows the device firmware to be updated by putting a firmware
file in /lib/firmware and providing the name of the file via the update_fw
sysfs property. The driver will then flash the firmware image into the
controller internal storage and restart the controller to activate the new
firmware.

The implementation was done by looking at the the messages passed between
the controller and proprietary vendor update tool. Not every detail of the
protocol is well understood, so the implementation still has some "monkey
see, monkey do" parts, as far as they have been found to be required for
the update to succeed.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>