net: dsa: mv88e6xxx: Don't consider port mode when getting SERDES
The Marvell 6390X family of switches can reassign SERDES interfaces
from ports 9 and 10 to the lower ports, if ports 9 and 10 are not
using them. When a lower port has both a SERDES and an embedded PHY,
the switch will monitor link on the PHY and SERDES, and when the first
gets link, connect it to the MAC. At this point, the ports cmode is
updated to indicate either 1000Base-X, or PHY.

The code to determine if a port has a SERDES was looking at the cmode
of the port, as well as the configuration of ports 9/10. If the port
is not using 1000Base-X, it was considered to not having a SERDES.
Hence the SERDES was not powered up, the interrupt not requested, etc.
Normally you would not expect this to work, a classic chicken/egg. If
the SERDES is not powered, it cannot get link, and hence the port will
never be in 1000Base-X mode, and so it never is determined to have a
SERDES. However, it seems that after reset, the SERDES is power up by
default. And if there happens to be a link partner, the SERDES gets
link, and sets the cmode to 1000Base-X. The driver then configures the
port down, powers off the SERDES, etc, until the interface is
opened. However the switch stays with 1000Base-X. So when the
interface is opened, it is determined the interface has a SERDES, and
it works.

It appears all previous testing was done with an interface connected
to a running peer. And hence the chicken/egg was never noticed.

Solve this by considering a port to have a SERDES purely based on if
port 9/10 is not using it.

Reported-by: Chris Healy <Chris.Healy@zii.aero>
Signed-off-by: Andrew Lunn <andrew@lunn.ch>