net: dsa: mv88e6xxx: rename port cmode to cmode_requested
Device tree and phylink can request a port uses a specified mode, eg
1000BaseX, SGMII. This gets translated into the mv88e6xxx cmode. For
some ports, it is not possible to write to the cmode bitfield. It will
however change cmode if the SERDES/PCS established link.

Rename the port cmode variable to cmode_requested to indicate it is
what has been requested, but might not yet actually be so in the
hardware.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>