net: dsa: mv88e6xxx: Allow setting cmode for mv88e6390X lower ports.
The mv88e6390X lower ports can make use of a SERDES if available.
When doing this, they use 1000BaseX. Allow the cmode to be set to
this. The hardware does not in fact allow the cmode to be set, but the
driver needs to make some decision based on if the port has been
requested to be in 1000BaseX mode, e.g. power up the SERDES, enable
interrupts etc.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>