net: dsa: mv88e6xxx: 88e6390X: Initialize port 9/10 is a slow mode
Ports 9 and 10 of the 88e6390X have 4 SERDES lanes. These lanes can be
used by ports 2-8 if they are not being used by ports 9 or 10. However
there is an ordering issue. Depending on strapping, ports 9 or 10
might default to a high speed mode, requiring all the SERDES lines.
Normally, the configuration of the port is only set when the interface
is brought up. But if a lower port, which is expected to have a SERDES
lane, is configured up before port 9/10, it could find its SERDES lane
is not available.

Work around this during port setup during probe. All user ports are
configured:

	err = mv88e6xxx_port_setup_mac(chip, port, LINK_UNFORCED,
				       SPEED_UNFORCED, DUPLEX_UNFORCED,
				       PAUSE_ON,
				       PHY_INTERFACE_MODE_NA);

meaning they should go with auto-negotiation. For ports 9 and 10, key
in on the PHY_INTERFACE_MODE_NA. Use this to configure the port to
1000BaseX, which only uses one SERDES lane. When ports 9/10 are later
configured up, with the correct PHY interface mode, they can steel
back there SERDES interfaces if needed. But until then, lower ports
which are legitimately expecting a SERDES interface can be configured
up and have the SERDES interface available.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>