net: phy: marvell: optimise TDR data retrieval
The majority of the runtime for getting the TDR data is performing
MDIO operations. Reduce the number of MDIO operations by taking the
bus lock only once and only changing the page once, rather than
locking for every access and changing the page for every access. This
can reduce the time takes to 1/3 of the unoptimized time.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>