net: phy: Makefile: Enable -Wtype-limits
PHY registers are u16, due to the nature of the MDIO bus.  A common
error with PHY drivers is assigning the return value of phy_read() to
a u16. However, phy_read() can return a negative error code. The
compiler, with its default set of warnings, does not report this
problem. It is only noticed later, when static analysis tools are
used, or the code is compiled with a higher warning level, that the
problem comes to light.

Enable Wtype-limits by default for PHY drivers so developers catch
this problem earlier.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>