net: phy: Add internal implementation of phy_read_status()
As a step towards fixing the locking with phy_read_status(), add an
internal implementation, which assumes the phydev->lock is held.
Convert the internal users to this new function.

Although phy_init_eee() is an phylib API call, it is mostly used in
the update_link callback, which already has the phydev locked. So it
should not take the lock again.

phy_check_link_status() already checks that the lock is held, so
making it clear the internal version should be used.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>