net: phy: Make phy_read_status() perform locking
Most users of this API call don't take phydev->lock. So there is a
potential race condition in that phylib might be manipulating the
phydev->link status, duplex etc, at the same time as the MAC driver.
To avoid this, make phy_read_status() perform locking.

The lan78xx driver did take the lock, so this now needs removing to
avoid a deadlock.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>