net: phy: phy_init_eee: Add locking
phy_init_eee() can be called in two different contexts. The MAC driver
can call it during open, or as part of an ethtool call. In this
context, the PHY lock is not held. phy_init_eee() then needs to take
the PHY lock. The second context is in the phylib adjust_link
callback, of the phylink mac_link_up() callback. In this context the
lock is already held. Add a new API call, phy_adjust_eee() which
should be used in these two callbacks. phy_init_eee() now takes the
PHY lock.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>