net: phylink: Hold PHY mutex while calling mac_link_up callback
The mac_link_up callback in phylink is similar to the adjust_link
callback in phylib. It is guaranteed that the PHY is locked when the
adjust_link callback is called, so that the phydev members are
consistent. Have phylink also lock the PHY to give the same assurances
when calling mac_link_up(). However do to mutex ordering, the PHY
mutex must be held before the phylink mutex. So the lock needs to be
taken in the calling function, not directly in phylink_link_up().

Signed-off-by: Andrew Lunn <andrew@lunn.ch>