platform/x86: Add Zodiac Radio Access Point platform driver
The device uses a CompExpress module on a baseboard which has a
Marvell Ethernet switch and a number of Atheros wireless access point
chipsets. Add a platform driver, which matches on Zodiac custom BIOS
DMS strings and instantiates the needed drivers for the platform.

NOT FOR MAINLINE

platform: x86: zii-rap: Add TQ-Systems DMI Strings

Hardware still exists which have the TQ-Sysems DMI Strings. Add them
back to allow testing on such systems.

NOT FOR MAINLINE