ZII platform driver: Use driver_override for gpio-decoder
Platform device is bound to a driver using the device
name. gpio-decoder then uses the device name as the input name. As a
result the input name is the rather uninformative "gpio_decoder:.

When driver_override is set, this name is used to find the driver,
leaving the device name to be more informative.