dt-bindings: spi: Convert spi-orion.txt to DT schema
The orion SPI driver is normally used in a polled IO mode. It does
however have a second mode which memory maps some buffers. No board in
mainline uses this mode, and it is not easy to get the example to work
because the C preprosessor is not used. So drop the example.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>